Web app/requirements.txt
scikit-learn
numpy
Pillow
streamlit
